pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0,<2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "pippack"
version = "0.0.1"
description = "Implementation of Protein Invariant Point Packer (PIPPack)"
authors = ["Name <email@address>"]

readme = "README.md"
license = "GNU"
repository = "https://github.com/Kuhlman-Lab/PIPPack"
classifiers = [
    "Topic :: Scientific/Engineering :: Biochemistry",
    "Topic :: Scientific/Engineering :: Protein Engineering"
]

packages = [
    { include = "pippack" },
    { include = "pippack/**/*.py" },
]

[tool.poetry.dependencies]
python = ">=3.9"
numpy = "1.26.3"
pooch = "1.8.0"
joblib = "^1.3.2"
biopython = ">=1.81"
lightning = "^2.0"
lightning-cloud = "^0.5.64"

torch_geometric="^2.4.0"

torch = { version = "^2.0.1", allow-prereleases = false}
torchvision = { version = "*", allow-prereleases = false}
torchaudio = { version = "*", allow-prereleases = false}

hydra-core= "*"
omegaconf = "*"

